refactor(infra): move mailbox decoder logic
